+++ src/messageExtension/MessageExtensionGenerator.ts
@@ -397,7 +397,7 @@ export class MessageExtensionGenerator extends Generator {
                     ], this.fs);
                 }
                 Yotilities.addAdditionalDeps([
-                    ["botbuilder-teams-messagingextensions", "1.3.0"]
+                    ["botbuilder-teams-messagingextensions", "1.4.0"]
                 ], this.fs);
 
                 if (this.options.messagingExtensionCanUpdateConfiguration || this.options.messagingExtensionActionResponseTypeConfig) {